requirements.txt
--extra-index-url https://rdil.mycloudrepo.io/public/repositories/cakebot

# development dependencies
black==19.10b0
mypy==0.770
flake8==3.7.9
ptpython==3.0.1
isort==4.3.21

# primary dependencies
discord.py==1.3.2
sentry-sdk==0.14.3
discord-sentry-reporting==2.0.4
sqlalchemy==1.3.15

# data libraries
iss-api==1.1
filehandlers==2.12.1
slots==1.3
factdata==2.0.0
area4>=3.0.0
reverse_geocoder==1.5.1
PyGithub==1.47

# extensions for better support
cchardet==2.1.6

# tests
unittest-xml-reporting==3.0.2
